refactor(e2e): #220 move federated harness container infra helpers
Pure move of startContainers/startDuckDB/createS3Client/cleanupContainers
from harness.go (501 lines, one over the cap) into harness_infra.go.
Byte-identity verified against HEAD. No behavior change.
